Fixed AnnotationInspectorTest to work on Java 9+
The `AnnotationInspectorTest` works by counting how many `@Deprecated` the
`AnnotationInspector` will find on a test class. This does not work anymore
because in Java 9, the method `Object.finalize()` gets deprecated as well.

While there is the possibility to if/else around this using class header magic,
this will bite us again when reflective access to `Object` will be forbidden.
Instead, I simply replaced `@Deprecated` with `@ExposeMethodResult`.